Split Compiler class into smaller ones
