+++ src/Jade/Compiler.php
@@ -2,10 +2,12 @@
 
 namespace Jade;
 
+use Jade\Compiler\MixinVisitor;
+
 /**
  * Class Jade Compiler.
  */
-class Compiler
+class Compiler extends MixinVisitor
 {
     /**
      * @const string
@@ -223,19 +225,6 @@ class Compiler
         }
         // Remove the $ wich are not needed
         return $code;
-    }
-
-    /**
-     * @param Nodes\Node $node
-     *
-     * @return array
-     */
-    public function visit(Nodes\Node $node)
-    {
-        // TODO: set debugging info
-        $this->visitNode($node);
-
-        return $this->buffer;
     }
 
     /**
@@ -850,122 +839,6 @@ class Compiler
         return $this->createPhpBlock($code);
     }
 
-    /**
-     * @param Nodes\Node $node
-     *
-     * @return mixed
-     */
-    protected function visitNode(Nodes\Node $node)
-    {
-        $fqn = get_class($node);
-        $parts = explode('\\', $fqn);
-        $name = end($parts);
-        $method = 'visit' . ucfirst(strtolower($name));
-
-        return $this->$method($node);
-    }
-
-    /**
-     * @param Nodes\CaseNode $node
-     */
-    protected function visitCasenode(Nodes\CaseNode $node)
-    {
-        $within = $this->withinCase;
-        $this->withinCase = true;
-
-        $this->switchNode = $node;
-        $this->visit($node->block);
-
-        if (!isset($this->switchNode)) {
-            unset($this->switchNode);
-            $this->indents--;
-
-            $code = $this->createCode('}');
-            $this->buffer($code);
-        }
-        $this->withinCase = $within;
-    }
-
-    /**
-     * @param Nodes\When $node
-     */
-    protected function visitWhen(Nodes\When $node)
-    {
-        $code = '';
-        $arguments = array();
-
-        if (isset($this->switchNode)) {
-            $code .= 'switch (%s) {';
-            $arguments[] = $this->switchNode->expr;
-            unset($this->switchNode);
-
-            $this->indents++;
-        }
-        if ('default' == $node->expr) {
-            $code .= 'default:';
-        } else {
-            $code .= 'case %s:';
-            $arguments[] = $node->expr;
-        }
-
-        array_unshift($arguments, $code);
-
-        $code = call_user_func_array(array($this, 'createCode'), $arguments);
-
-        $this->buffer($code);
-
-        $this->visit($node->block);
-
-        $code = $this->createCode('break;');
-        $this->buffer($code . $this->newline());
-    }
-
-    /**
-     * @param Nodes\Literal $node
-     */
-    protected function visitLiteral(Nodes\Literal $node)
-    {
-        $str = preg_replace('/\\n/', '\\\\n', $node->string);
-        $this->buffer($str);
-    }
-
-    /**
-     * @param Nodes\Block $block
-     */
-    protected function visitBlock(Nodes\Block $block)
-    {
-        foreach ($block->nodes as $n) {
-            $this->visit($n);
-        }
-    }
-
-    /**
-     * @param Nodes\Doctype $doctype
-     *
-     * @throws \Exception
-     */
-    protected function visitDoctype(Nodes\Doctype $doctype = null)
-    {
-        if (isset($this->hasCompiledDoctype)) {
-            throw new \Exception('Revisiting doctype');
-        }
-        $this->hasCompiledDoctype = true;
-
-        $doc = (empty($doctype->value) || $doctype == null || !isset($doctype->value))
-            ? 'default'
-            : strtolower($doctype->value);
-
-        $str = isset($this->doctypes[$doc])
-            ? $this->doctypes[$doc]
-            : "<!DOCTYPE {$doc}>";
-
-        $this->buffer($str . $this->newline());
-
-        $this->terse = (strtolower($str) == '<!doctype html>');
-
-        $this->xml = ($doc == 'xml');
-    }
-
     protected static function initArgToNull(&$arg)
     {
         $arg = static::addDollarIfNeeded(trim($arg));
@@ -1031,404 +904,6 @@ class Compiler
     }
 
     /**
-     * @param Nodes\Mixin $mixin
-     */
-    protected function visitMixin(Nodes\Mixin $mixin)
-    {
-        $name = strtr($mixin->name, '-', '_') . '_mixin';
-        $blockName = var_export($mixin->name, true);
-        if ($this->allowMixinOverride) {
-            $name = '$GLOBALS[\'' . $name . '\']';
-        }
-        $arguments = $mixin->arguments;
-        $block = $mixin->block;
-        $attributes = static::decodeAttributes($mixin->attributes);
-
-        if ($mixin->call) {
-            $defaultAttributes = array();
-            $newArrayKey = null;
-            $arguments = explode(',', $arguments);
-            $containsOnlyArrays = true;
-            foreach ($arguments as $key => &$argument) {
-                if (preg_match('`^\s*[a-zA-Z][a-zA-Z0-9:_-]*\s*=`', $argument)) {
-                    $tab = explode('=', trim($argument), 2);
-                    if (is_null($newArrayKey)) {
-                        $newArrayKey = $key;
-                        $argument = array();
-                    } else {
-                        unset($arguments[$key]);
-                    }
-                    if (count($tab) === 2) {
-                        $defaultAttributes[] = var_export($tab[0], true) . ' => ' . $tab[1];
-                        $arguments[$newArrayKey][$tab[0]] = static::decodeValue($tab[1]);
-                    } else {
-                        $arguments[$newArrayKey][$tab[0]] = true;
-                    }
-                } else {
-                    $containsOnlyArrays = false;
-                    $newArrayKey = null;
-                }
-            }
-            $arguments = array_map(function ($argument) {
-                if (is_array($argument)) {
-                    $argument = var_export($argument, true);
-                }
-
-                return $argument;
-            }, $arguments);
-
-            $defaultAttributes = implode(', ', $defaultAttributes);
-            if (!count($attributes)) {
-                $attributes = "(isset(\$attributes)) ? \$attributes : array($defaultAttributes)";
-            } else {
-                $_attr = array();
-                foreach ($attributes as $data) {
-                    if ($data['value'] === 'null' || $data['value'] === 'undefined' || is_null($data['value'])) {
-                        $_attr[$data['name']] = null;
-                    } elseif ($data['value'] === 'false' || is_bool($data['value'])) {
-                        $_attr[$data['name']] = false;
-                    } else {
-                        $value = trim($data['value']);
-                        if (strlen($value) && false !== strpos('\'"', $quote = substr($value, 0, 1))) {
-                            $data['value'] = stripslashes(substr($value, 1, -1));
-                        }
-                        $_attr[$data['name']] = $data['escaped'] === true
-                            ? htmlspecialchars($data['value'])
-                            : $data['value'];
-                    }
-                }
-
-                $attributes = var_export($_attr, true);
-                $mixinAttributes = var_export(static::decodeAttributes($mixin->attributes), true);
-                $attributes = "array_merge(\\Jade\\Compiler::withMixinAttributes($attributes, $mixinAttributes), (isset(\$attributes)) ? \$attributes : array($defaultAttributes))";
-            }
-
-            if ($block) {
-                $code = $this->createCode("\\Jade\\Compiler::recordMixinBlock($blockName, function (\$attributes) {");
-                $this->buffer($code);
-                $this->visit($block);
-                $this->buffer($this->createCode('});'));
-            }
-
-            if ($arguments === false || $arguments === null) {
-                $code = $this->createPhpBlock("{$name}({$attributes})");
-            } else {
-                if ($mixin->call) {
-                    $strings = array();
-                    $arguments = preg_replace_callback(
-                        '#([\'"])(.*(?!<\\\\)(?:\\\\{2})*)\\1#U',
-                        function ($match) use (&$strings) {
-                            $id = count($strings);
-                            $strings[] = $match[0];
-
-                            return 'stringToReplaceBy' . $id . 'ThCapture';
-                        },
-                        $arguments
-                    );
-                    $arguments = array_map(
-                        function ($arg) use ($strings) {
-                            return preg_replace_callback(
-                                '#stringToReplaceBy([0-9]+)ThCapture#',
-                                function ($match) use ($strings) {
-                                    return $strings[intval($match[1])];
-                                },
-                                $arg
-                            );
-                        },
-                        $arguments
-                    );
-                }
-
-                array_unshift($arguments, $attributes);
-                $arguments = array_filter($arguments, 'strlen');
-                $statements = $this->apply('createStatements', $arguments);
-
-                $variables = array_pop($statements);
-                if ($mixin->call && $containsOnlyArrays) {
-                    array_splice($variables, 1, 0, array('null'));
-                }
-                $variables = implode(', ', $variables);
-                array_push($statements, $variables);
-
-                $arguments = $statements;
-
-                $code_format = str_repeat('%s;', count($arguments) - 1) . "{$name}(%s)";
-
-                array_unshift($arguments, $code_format);
-
-                $code = $this->apply('createCode', $arguments);
-            }
-            $this->buffer($code);
-            if ($block) {
-                $code = $this->createCode("\\Jade\\Compiler::terminateMixinBlock($blockName);");
-                $this->buffer($code);
-            }
-        } else {
-            $previousVisitedMixin = isset($this->visitedMixin) ? $this->visitedMixin : null;
-            $this->visitedMixin = $mixin;
-            if ($arguments === null || empty($arguments)) {
-                $arguments = array();
-            } elseif (!is_array($arguments)) {
-                $arguments = array($arguments);
-            }
-
-            array_unshift($arguments, 'attributes');
-            $arguments = implode(',', $arguments);
-            $arguments = explode(',', $arguments);
-            array_walk($arguments, array(get_class(), 'initArgToNull'));
-            if ($this->allowMixinOverride) {
-                $code = $this->createCode("{$name} = function (%s) { ", implode(',', $arguments));
-
-                $this->buffer($code);
-                $this->indents++;
-                $this->visit($block);
-                $this->indents--;
-                $this->buffer($this->createCode('};'));
-            } else {
-                $code = $this->createCode("if(!function_exists('{$name}')) { function {$name}(%s) {", implode(',', $arguments));
-
-                $this->buffer($code);
-                $this->indents++;
-                $this->visit($block);
-                $this->indents--;
-                $this->buffer($this->createCode('} }'));
-            }
-            if (is_null($previousVisitedMixin)) {
-                unset($this->visitedMixin);
-            } else {
-                $this->visitedMixin = $previousVisitedMixin;
-            }
-        }
-    }
-
-    /**
-     * @param Nodes\Mixin $mixin
-     */
-    protected function visitMixinBlock(Nodes\MixinBlock $mixinBlock)
-    {
-        $name = var_export($this->visitedMixin->name, true);
-        $code = $this->createCode("\\Jade\\Compiler::callMixinBlock($name, \$attributes);");
-
-        $this->buffer($code);
-    }
-
-    /**
-     * @param Nodes\Tag $tag
-     */
-    protected function visitTag(Nodes\Tag $tag)
-    {
-        if (isset($tag->buffer)) {
-            if (preg_match('`^[a-z][a-zA-Z0-9]+(?!\()`', $tag->name)) {
-                $tag->name = '$' . $tag->name;
-            }
-            $tag->name = trim($this->createCode('echo ' . $tag->name . ';'));
-        }
-        if (!isset($this->hasCompiledDoctype) && 'html' == $tag->name) {
-            $this->visitDoctype();
-        }
-
-        $self_closing = (in_array(strtolower($tag->name), $this->selfClosing) || $tag->selfClosing) && !$this->xml;
-
-        if ($tag->name == 'pre') {
-            $pp = $this->prettyprint;
-            $this->prettyprint = false;
-        }
-
-        $noSlash = (!$self_closing || $this->terse);
-
-        if (count($tag->attributes)) {
-            $open = '<' . $tag->name;
-            $close = $noSlash ? '>' : ' />';
-
-            $this->buffer($this->indent() . $open, false);
-            $this->visitAttributes($tag->attributes);
-            $this->buffer($close . $this->newline(), false);
-        } else {
-            $html_tag = '<' . $tag->name . ($noSlash ? '>' : ' />');
-
-            $this->buffer($html_tag);
-        }
-
-        if (!$self_closing) {
-            $this->indents++;
-            if (isset($tag->code)) {
-                $this->visitCode($tag->code);
-            }
-            $this->visit($tag->block);
-            $this->indents--;
-
-            $this->buffer('</' . $tag->name . '>');
-        }
-
-        if ($tag->name == 'pre') {
-            $this->prettyprint = $pp;
-        }
-    }
-
-    /**
-     * @param Nodes\Filter $node
-     *
-     * @throws \InvalidArgumentException
-     */
-    protected function visitFilter(Nodes\Filter $node)
-    {
-        // Check that filter is registered
-        if (!array_key_exists($node->name, $this->filters)) {
-            throw new \InvalidArgumentException($node->name . ': Filter doesn\'t exists');
-        }
-
-        $filter = $this->filters[$node->name];
-
-        // Filters can be either a iFilter implementation, nor a callable
-        if (is_string($filter)) {
-            $filter = new $filter();
-        }
-        if (!is_callable($filter)) {
-            throw new \InvalidArgumentException($node->name . ': Filter must be callable');
-        }
-        $this->buffer($filter($node, $this));
-    }
-
-    /**
-     * @param Nodes\Text $text
-     */
-    protected function visitText(Nodes\Text $text)
-    {
-        $this->buffer($this->interpolate($text->value));
-    }
-
-    /**
-     * @param Nodes\Comment $comment
-     */
-    protected function visitComment(Nodes\Comment $comment)
-    {
-        if ($comment->buffer) {
-            $this->buffer('<!--' . $comment->value . '-->');
-        }
-    }
-
-    /**
-     * @param Nodes\BlockComment $comment
-     */
-    protected function visitBlockComment(Nodes\BlockComment $comment)
-    {
-        if (!$comment->buffer) {
-            return;
-        }
-
-        if (strlen($comment->value) && 0 === strpos(trim($comment->value), 'if')) {
-            $this->buffer('<!--[' . trim($comment->value) . ']>');
-            $this->visit($comment->block);
-            $this->buffer('<![endif]-->');
-        } else {
-            $this->buffer('<!--' . $comment->value);
-            $this->visit($comment->block);
-            $this->buffer('-->');
-        }
-    }
-
-    /**
-     * @param Nodes\Code $node
-     */
-    protected function visitCode(Nodes\Code $node)
-    {
-        $code = trim($node->value);
-
-        if ($node->buffer) {
-            $pattern = $node->escape ? static::ESCAPED : static::UNESCAPED;
-            $this->buffer($this->createCode($pattern, $code));
-        } else {
-            $php_open = implode('|', $this->phpOpenBlock);
-
-            if (preg_match("/^[[:space:]]*({$php_open})(.*)/", $code, $matches)) {
-                $code = trim($matches[2], '; ');
-                while (($len = strlen($code)) > 1 && ($code[0] == '(' || $code[0] == '{') && ord($code[0]) == ord(substr($code, -1)) - 1) {
-                    $code = trim(substr($code, 1, $len - 2));
-                }
-
-                $index = count($this->buffer) - 1;
-                $conditional = '';
-
-                if (isset($this->buffer[$index]) && false !== strpos($this->buffer[$index], $this->createCode('}'))) {
-                    // the "else" statement needs to be in the php block that closes the if
-                    $this->buffer[$index] = null;
-                    $conditional .= '} ';
-                }
-
-                $conditional .= '%s';
-
-                if (strlen($code) > 0) {
-                    $conditional .= '(%s) {';
-                    if ($matches[1] == 'unless') {
-                        $conditional = sprintf($conditional, 'if', '!(%s)');
-                    } else {
-                        $conditional = sprintf($conditional, $matches[1], '%s');
-                    }
-                    $this->buffer($this->createCode($conditional, $code));
-                } else {
-                    $conditional .= ' {';
-                    $conditional = sprintf($conditional, $matches[1]);
-
-                    $this->buffer($this->createCode($conditional));
-                }
-            } else {
-                $this->buffer($this->createCode('%s', $code));
-            }
-        }
-
-        if (isset($node->block)) {
-            $this->indents++;
-            $this->visit($node->block);
-            $this->indents--;
-
-            if (!$node->buffer) {
-                $this->buffer($this->createCode('}'));
-            }
-        }
-    }
-
-    /**
-     * @param $node
-     */
-    protected function visitEach($node)
-    {
-
-        //if (is_numeric($node->obj)) {
-        //if (is_string($node->obj)) {
-        //$serialized = serialize($node->obj);
-        if (isset($node->alternative)) {
-            $code = $this->createCode('if (isset(%s) && %s) {', $node->obj, $node->obj);
-            $this->buffer($code);
-            $this->indents++;
-        }
-
-        if (isset($node->key) && mb_strlen($node->key) > 0) {
-            $code = $this->createCode('foreach (%s as %s => %s) {', $node->obj, $node->key, $node->value);
-        } else {
-            $code = $this->createCode('foreach (%s as %s) {', $node->obj, $node->value);
-        }
-
-        $this->buffer($code);
-
-        $this->indents++;
-        $this->visit($node->block);
-        $this->indents--;
-
-        $this->buffer($this->createCode('}'));
-
-        if (isset($node->alternative)) {
-            $this->indents--;
-            $this->buffer($this->createCode('} else {'));
-            $this->indents++;
-
-            $this->visit($node->alternative);
-            $this->indents--;
-
-            $this->buffer($this->createCode('}'));
-        }
-    }
-
-    /**
      * @param $attributes
      */
     public static function displayAttributes($attributes, $quote)
@@ -1450,106 +925,5 @@ class Compiler
     public static function isDisplayable($value)
     {
         return !is_null($value) && $value !== false;
-    }
-
-    /**
-     * @param $attributes
-     */
-    protected function visitAttributes($attributes)
-    {
-        $pp = $this->prettyprint;
-        $this->prettyprint = false;
-        $items = array();
-        $classes = array();
-        $classesCheck = array();
-        $quote = var_export($this->quote, true);
-
-        foreach ($attributes as $attr) {
-            $key = trim($attr['name']);
-            if ($key === '&attributes') {
-                $addClasses = '';
-                if (count($classes) || count($classesCheck)) {
-                    foreach ($classes as &$value) {
-                        $value = var_export($value, true);
-                    }
-                    foreach ($classesCheck as $value) {
-                        $statements = $this->createStatements($value);
-                        $classes[] = $statements[0][0];
-                    }
-                    $addClasses = '$__attributes["class"] = ' .
-                        'implode(" ", array(' . implode(', ', $classes) . ')) . ' .
-                        '(empty($__attributes["class"]) ? "" : " " . $__attributes["class"]); ';
-                    $classes = array();
-                    $classesCheck = array();
-                }
-                $value = empty($attr['value']) ? 'attributes' : $attr['value'];
-                $statements = $this->createStatements($value);
-                $items[] = $this->createCode(
-                    '$__attributes = ' . $statements[0][0] . ';' .
-                    $addClasses .
-                    '\\Jade\\Compiler::displayAttributes($__attributes, ' . $quote . ');');
-            } else {
-                $valueCheck = null;
-                $value = trim($attr['value']);
-
-                if ($this->isConstant($value, $key == 'class')) {
-                    $value = trim($value, ' \'"');
-                    if ($value === 'undefined') {
-                        $value = 'null';
-                    }
-                } else {
-                    $json = static::parseValue($value);
-
-                    if ($json !== null && is_array($json) && $key == 'class') {
-                        $value = implode(' ', $json);
-                    } elseif ($key == 'class') {
-                        if ($this->keepNullAttributes) {
-                            $value = $this->createCode('echo (is_array($_a = %1$s)) ? implode(" ", $_a) : $_a', $value);
-                        } else {
-                            $statements = $this->createStatements($value);
-                            $classesCheck[] = '(is_array($_a = ' . $statements[0][0] . ') ? implode(" ", $_a) : $_a)';
-                            $value = 'null';
-                        }
-                    } elseif ($this->keepNullAttributes) {
-                        $value = $this->createCode(static::UNESCAPED, $value);
-                    } else {
-                        $valueCheck = $value;
-                        $value = $this->createCode(static::UNESCAPED, '$__value');
-                    }
-                }
-
-                if ($key == 'class') {
-                    if ($value !== 'false' && $value !== 'null' && $value !== 'undefined') {
-                        array_push($classes, $value);
-                    }
-                } elseif ($value == 'true' || $attr['value'] === true) {
-                    $items[] = ' ' . $key . ($this->terse
-                        ? ''
-                        : '=' . $this->quote . $key . $this->quote
-                    );
-                } elseif ($value !== 'false' && $value !== 'null' && $value !== 'undefined') {
-                    $items[] = is_null($valueCheck)
-                        ? ' ' . $key . '=' . $this->quote . $value . $this->quote
-                        : $this->createCode('if (\\Jade\\Compiler::isDisplayable($__value = %1$s)) { ', $valueCheck)
-                            . ' ' . $key . '=' . $this->quote . $value . $this->quote
-                            . $this->createCode('}');
-                }
-            }
-        }
-
-        if (count($classes)) {
-            if (count($classesCheck)) {
-                $classes[] = $this->createCode('echo implode(" ", array(' . implode(', ', $classesCheck) . '))');
-            }
-            $items[] = ' class=' . $this->quote . implode(' ', $classes) . $this->quote;
-        } elseif (count($classesCheck)) {
-            $item = $this->createCode('if("" !== ($__classes = implode(" ", array(' . implode(', ', $classesCheck) . ')))) {');
-            $item .= ' class=' . $this->quote . $this->createCode('echo $__classes') . $this->quote;
-            $items[] = $item . $this->createCode('}');
-        }
-
-        $this->prettyprint = $pp;
-
-        $this->buffer(' ' . trim(implode('', $items)), false);
     }
 }
